ajout token identification

+++ suppression-article.php
@@ -4,14 +4,22 @@ include 'header-json.php';
 
 try{
 
-    include 'bdd.php';
+    include 'authentification.php';
 
-    $requete = $bdd->prepare("DELETE FROM article WHERE id = :id");
+    if($utilisateurConnecte) {
 
-    $requete->execute(["id" => $_GET['id']]);
+        $requete = $bdd->prepare("DELETE FROM article WHERE id = :id");
 
-   // echo '{"message" => "L\'article a bien été supprimé"}';
-   echo json_encode(["message" => "L'article a bien été supprimé"]);
+        $requete->execute(["id" => $_GET['id']]);
+
+        // echo '{"message" => "L\'article a bien été supprimé"}';
+        echo json_encode(["message" => "L'article a bien été supprimé"]);
+
+    } else {
+
+        echo json_encode(["message" => "Ce compte n'existe plus"]);
+
+    }
 
 } catch (PDOException $e) {
     echo 'Echec de la connexion : ' . $e->getMessage();
